Move BankDetails to src/components/*
