+++ src/components/BankDetails/RecipientDetails/helpers/dot.ts
