+++ src/components/BankDetails/RecipientDetails/helpers/isTextType.ts
